feat: enhance Kling request handling and response structure
- Added logic to determine task types based on the original request path, improving routing for multi-modal video editing endpoints.
- Set default model names for auxiliary interfaces that do not require a model parameter, ensuring proper channel routing.
- Updated the `OriginalDuration` field in the response payload to use a float type for better precision in video duration representation.

+++ relay/channel/task/kling/adaptor.go
@@ -154,7 +154,7 @@ type multiElementsInitResponsePayload struct {
 		Status           int     `json:"status"`            // 拒识码，非0为识别失败
 		SessionId        string  `json:"session_id"`        // 会话ID，有效期24小时
 		Fps              float64 `json:"fps"`               // 解析后视频的帧数
-		OriginalDuration int     `json:"original_duration"` // 解析后视频的时长
+		OriginalDuration float64 `json:"original_duration"` // 解析后视频的时长（毫秒）
 		Width            int     `json:"width"`             // 解析后视频的宽
 		Height           int     `json:"height"`            // 解析后视频的高
 		TotalFrame       int     `json:"total_frame"`       // 解析后视频的总帧数
